requirements.txt
jax>=0.4.30
jaxlib>=0.4.30
grain-nightly==0.0.10
google-cloud-storage==2.17.0
absl-py
datasets
flax>=0.10.2
optax>=0.2.3
torch==2.5.1
torchvision==0.20.1
ftfy
tensorboard>=2.17.0
tensorboardx==2.6.2.2
tensorboard-plugin-profile==2.15.2
Jinja2
scikit-image
parameterized
Pillow
pylint
pyink
pytest==8.2.2
tensorflow==2.17.0
tensorflow-datasets>=4.9.6
ruff>=0.1.5,<=0.2
git+https://github.com/mlperf/logging.git
opencv-python-headless==4.10.0.84
orbax-checkpoint==0.10.3
tokenizers==0.21.0
huggingface_hub==0.24.7
transformers==4.48.1
einops==0.8.0
sentencepiece
aqtp
audax
jieba
pypinyin
librosa
jax_vocos